cyrisk/requirements.txt
aiohttp==3.8.3
aioresponses==0.7.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.1
certifi==2022.12.7
channels==4.0.0
charset-normalizer==2.1.1
Django==3.2.16
django-background-tasks==1.2.5
django-channels==0.7.0
django-compat==1.0.15
django-redis==5.2.0
djangorestframework==3.14.0
frozenlist==1.3.3
httpretty==1.1.4
idna==3.4
lxml==4.9.2
multidict==6.0.4
oauthlib==3.2.2
python-Wappalyzer==0.3.1
pytz==2022.7
redis==4.4.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.13
websockets==10.4
yarl==1.8.2
